requirements.txt
dnspython : 2.6.1
